+++ docker/cron/cron-execution_gfs.sh
@@ -4,4 +4,4 @@ scriptPath=$(dirname "$(readlink -f "$0")")
 source "${scriptPath}/.env.sh"
 
 # Code of python:3.12-slim is located under /usr/local/bin/python3
-/usr/local/bin/python3 /app/src/ecmwf_download.py "$@"
+/usr/local/bin/python3 /app/src/gfs_fc_engine.py "$@"