+++ CHANGELOG.md
@@ -6,10 +6,19 @@
 ### Deprecated
 ### Removed
 ### Security
+## 1.1.0 (2025-02-24)
+### Added
+Energy prediction for the PV installation augmented
+by utilizing the GFS weather forecast of NCEP (NOAA)
+### Changed
+### Fixed
+### Deprecated
+### Removed
+### Security
 ## 1.0.1 (2025-01-12)
 ### Added
 ### Changed
-- SunInflux calculation optimized for varying intervals
+- Class SunInflux calculation optimized for varying intervals
 ### Fixed
 ### Deprecated
 ### Removed